+++ src/js/helper/domHelper.js
@@ -7,7 +7,7 @@ export default class DomHelper {
      * @param {?string} text
      * @returns {HTMLElement}
      */
-    createElement(name, attributes, text) {
+    static createElement(name, attributes, text) {
         const node = document.createElement(name);
 
         if ('undefined' !== typeof attributes) {
@@ -23,5 +23,27 @@ export default class DomHelper {
         }
 
         return node;
+    }
+
+    /**
+     * Hide a jQuery wrapped DOM element
+     *
+     * @param {jQuery} $element
+     */
+    static hideElement($element) {
+        if ($element.css('display') === 'block') {
+            $element.hide();
+        }
+    }
+
+    /**
+     * Show a jQuery wrapped DOM element
+     *
+     * @param {jQuery} $element
+     */
+    static showElement($element) {
+        if ($element.css('display') !== 'block') {
+            $element.show();
+        }
     }
 }