+++ src/js/item/audioItem.js
@@ -1,4 +1,5 @@
 import BaseItem from "./baseItem";
+import DomHelper from "../helper/domHelper";
 import FileHelper from "./../helper/fileHelper";
 
 /**
@@ -6,11 +7,11 @@ import FileHelper from "./../helper/fileHelper";
  */
 export default class AudioItem extends BaseItem {
     /**
-     * Allowed file extensions for audio tag
+     * Allowed file extensions
      *
      * @returns {{mp3: string, wav: string, ogg: string}}
      */
-    static fileFormats() {
+    static supportedFileFormats() {
         return {
             'mp3': 'audio/mpeg',
             'ogg': 'audio/ogg',
@@ -35,25 +36,22 @@ export default class AudioItem extends BaseItem {
 
         this.path = parameters.path;
         this.caption = parameters.caption;
-
         this.fileExtension = FileHelper.guessExtension(this.path);
 
-        if (!AudioItem.fileFormats().hasOwnProperty(this.fileExtension)) {
-            throw Error('Unsupported file extension "' + this.fileExtension + '"');
-        }
+        FileHelper.checkFileFormat(this.fileExtension, AudioItem.supportedFileFormats());
     }
 
     /**
      * @returns {HTMLElement}
      */
     createAudio() {
-        const audio = this.domHelper.createElement('audio', {'class': 'genuine-theme'}, AudioItem.unsupportedTagMessage());
+        const audio = DomHelper.createElement('audio', {'class': 'genuine-theme'}, AudioItem.unsupportedTagMessage());
         audio.setAttribute('controls', '');
         audio.setAttribute('preload', 'metadata');
 
-        const source = this.domHelper.createElement('source');
+        const source = DomHelper.createElement('source');
         source.setAttribute('src', this.path);
-        source.setAttribute('type', AudioItem.fileFormats()[this.fileExtension]);
+        source.setAttribute('type', AudioItem.supportedFileFormats()[this.fileExtension]);
 
         audio.appendChild(source);
 
@@ -65,14 +63,14 @@ export default class AudioItem extends BaseItem {
      */
     renderHtml() {
         const element = this.createItemElement();
-        const audioItem = this.domHelper.createElement('div', {'class': 'audio-item'});
-
-        if ('undefined' !== typeof this.caption) {
-            const caption = this.domHelper.createElement('span', {'class': 'caption'}, this.caption);
-            audioItem.appendChild(caption);
-        }
+        const audioItem = DomHelper.createElement('div', {'class': 'audio-item'});
 
         audioItem.appendChild(this.createAudio());
+
+        if ('undefined' !== typeof this.caption) {
+            const caption = DomHelper.createElement('span', {'class': 'caption'}, this.caption);
+            audioItem.appendChild(caption);
+        }
 
         element.appendChild(audioItem);