+++ src/js/item/baseItem.js
@@ -6,10 +6,9 @@ export default class BaseItem {
      * @param {object} parameters
      */
     constructor(parameters) {
-        this.domHelper = new DomHelper();
-
         this.identifier = UniqueId.generate('item');
         this.position = typeof parameters.position !== 'undefined' ? parameters.position : {left: 0, top: 0};
+        this.sticky = parameters.sticky;
     }
 
     checkRequiredParameters(parameters, requiredParameters) {
@@ -24,7 +23,7 @@ export default class BaseItem {
      * @returns {HTMLElement}
      */
     createHotspotElement() {
-        const element = this.domHelper.createElement('div', {'class': 'hotspot icon-radio-checked'});
+        const element = DomHelper.createElement('div', {'class': 'hotspot icon-radio-checked'});
         element.setAttribute('data-for', this.identifier);
         element.style.left = this.position.left + 'px';
         element.style.top = this.position.top + 'px';
@@ -36,8 +35,18 @@ export default class BaseItem {
      * @returns {HTMLElement}
      */
     createItemElement() {
-        const element = this.domHelper.createElement('div', {'class': 'item'});
+        let itemClass = 'item';
+        if (this.sticky === true) {
+            itemClass += ' behavior-sticky';
+        }
+
+        const element = DomHelper.createElement('div', {'class': itemClass});
         element.setAttribute('data-id', this.identifier);
+
+        if (this.sticky === true) {
+            const closeButton = DomHelper.createElement('div', {'class': 'close-button icon-cancel-circle'});
+            element.appendChild(closeButton);
+        }
 
         return element;
     }