+++ src/js/item/videoItem.js
@@ -1,4 +1,5 @@
 import BaseItem from "./baseItem";
+import DomHelper from "../helper/domHelper";
 import FileHelper from "./../helper/fileHelper";
 
 /**
@@ -6,11 +7,11 @@ import FileHelper from "./../helper/fileHelper";
  */
 export default class VideoItem extends BaseItem {
     /**
-     * Allowed file extensions for video tag
+     * Allowed file extensions
      *
      * @returns {{mp4: string, webm: string}}
      */
-    static fileFormats() {
+    static supportedFileFormats() {
         return {
             'mp4': 'video/mp4',
             'webm': 'video/webm'
@@ -35,26 +36,23 @@ export default class VideoItem extends BaseItem {
         this.path = parameters.path;
         this.caption = parameters.caption;
         this.poster = parameters.poster;
-
         this.fileExtension = FileHelper.guessExtension(this.path);
 
-        if (!VideoItem.fileFormats().hasOwnProperty(this.fileExtension)) {
-            throw Error('Unsupported file extension "' + this.fileExtension + '"');
-        }
+        FileHelper.checkFileFormat(this.fileExtension, VideoItem.supportedFileFormats());
     }
 
     /**
      * @returns {HTMLElement}
      */
     createVideo() {
-        const video = this.domHelper.createElement('video', {'class': 'genuine-theme'}, VideoItem.unsupportedTagMessage());
+        const video = DomHelper.createElement('video', {'class': 'genuine-theme'}, VideoItem.unsupportedTagMessage());
         video.setAttribute('controls', '');
         video.setAttribute('controlsList', 'nodownload');
         video.setAttribute('preload', 'metadata');
 
-        const source = this.domHelper.createElement('source');
+        const source = DomHelper.createElement('source');
         source.setAttribute('src', this.path);
-        source.setAttribute('type', VideoItem.fileFormats()[this.fileExtension]);
+        source.setAttribute('type', VideoItem.supportedFileFormats()[this.fileExtension]);
 
         if ('undefined' !== typeof this.poster) {
             video.setAttribute('poster', this.poster);
@@ -70,14 +68,14 @@ export default class VideoItem extends BaseItem {
      */
     renderHtml() {
         const element = this.createItemElement();
-        const videoItem = this.domHelper.createElement('div', {'class': 'video-item'});
-
-        if ('undefined' !== typeof this.caption) {
-            const caption = this.domHelper.createElement('span', {'class': 'caption'}, this.caption);
-            videoItem.appendChild(caption);
-        }
+        const videoItem = DomHelper.createElement('div', {'class': 'video-item'});
 
         videoItem.appendChild(this.createVideo());
+
+        if ('undefined' !== typeof this.caption) {
+            const caption = DomHelper.createElement('span', {'class': 'caption'}, this.caption);
+            videoItem.appendChild(caption);
+        }
 
         element.appendChild(videoItem);