+++ test/item/audioItemTest.js
@@ -25,22 +25,22 @@ describe('Audio Item', function() {
             expect(audioContainerElement.nodeName).to.equal('DIV');
             expect(audioContainerElement.getAttribute('class')).to.equal('audio-item');
 
-            let captionElement = audioContainerElement.childNodes[0];
-            expect(captionElement.nodeName).to.equal('SPAN');
-            expect(captionElement.getAttribute('class')).to.equal('caption');
-            expect(captionElement.textContent).to.equal('my test caption');
-
-            let audioElement = audioContainerElement.childNodes[1];
+            let audioElement = audioContainerElement.childNodes[0];
             expect(audioElement.nodeName).to.equal('AUDIO');
             expect(audioElement.getAttribute('class')).to.equal('genuine-theme');
             expect(audioElement.hasAttribute('controls')).to.be.true;
             expect(audioElement.getAttribute('preload')).to.equal('metadata');
-            expect(audioElement.textContent).to.equal('Your browser does not support the audio tag.');
+            expect(audioElement.textContent).to.equal('Your browser does not support the audio tag.')
 
             let sourceElement = audioElement.childNodes[1];
             expect(sourceElement.nodeName).to.equal('SOURCE');
             expect(sourceElement.getAttribute('src')).to.equal('test.mp3');
             expect(sourceElement.getAttribute('type')).to.equal('audio/mpeg');
+
+            let captionElement = audioContainerElement.childNodes[1];
+            expect(captionElement.nodeName).to.equal('SPAN');
+            expect(captionElement.getAttribute('class')).to.equal('caption');
+            expect(captionElement.textContent).to.equal('my test caption');
         });
     });
 });